requirements.txt
opencv-python
mediapipe==0.10.9
numpy
websocket-client
